working replacement for the qp solver. currently solves for all signs and returns all solutions
